Move time utils to utils folder, add 12h stup to clock
