+++ lib/src/DatePicker/Calendar.jsx
@@ -7,7 +7,7 @@ import { extendMoment } from 'moment-range';
 import classnames from 'classnames';
 import CalendarHeader from './CalendarHeader';
 import DomainPropTypes from '../constants/prop-types';
-import * as defaultUtils from '../_shared/utils';
+import * as defaultUtils from '../utils/utils';
 
 const moment = extendMoment(Moment);
