+++ lib/src/DatePicker/CalendarHeader.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles, IconButton } from 'material-ui';
-import * as defaultUtils from '../_shared/utils';
+import * as defaultUtils from '../utils/utils';
 
 export const CalendarHeader = (props) => {
   const {
