+++ lib/src/DatePicker/YearSelection.jsx
@@ -5,7 +5,7 @@ import classnames from 'classnames';
 import { extendMoment } from 'moment-range';
 import { withStyles } from 'material-ui';
 import DomainPropTypes from '../constants/prop-types';
-import * as defaultUtils from '../_shared/utils';
+import * as defaultUtils from '../utils/utils';
 
 const moment = extendMoment(Moment);
