+++ lib/src/TimePicker/TimePicker.jsx
@@ -5,8 +5,8 @@ import PickerToolbar from '../_shared/PickerToolbar';
 import ToolbarButton from '../_shared/ToolbarButton';
 import HourView from './HourView';
 import MinutesView from './MinutesView';
-import { convertToMeridiem } from './utils/time-utils';
-import * as defaultUtils from '../_shared/utils';
+import { convertToMeridiem } from '../utils/time-utils';
+import * as defaultUtils from '../utils/utils';
 
 export class TimePicker extends Component {
   static propTypes = {
@@ -16,11 +16,13 @@ export class TimePicker extends Component {
     theme: PropTypes.object.isRequired,
     children: PropTypes.node,
     utils: PropTypes.object,
+    ampm: PropTypes.bool,
   }
 
   static defaultProps = {
     children: null,
     utils: defaultUtils,
+    ampm: true,
   }
 
   state = {
@@ -60,7 +62,7 @@ export class TimePicker extends Component {
 
   render() {
     const {
-      classes, theme, date, utils,
+      classes, theme, date, utils, ampm,
     } = this.props;
 
     const { isHourViewShown, meridiemMode } = this.state;
@@ -125,6 +127,7 @@ export class TimePicker extends Component {
                 meridiemMode={meridiemMode}
                 onChange={this.handleChange(true)}
                 utils={utils}
+                ampm={ampm}
               />
             :
               <MinutesView
